obsługa usuwania wypozyczeń gdy nie ma wypozyczen

+++ biblioteka/apka/views.py
@@ -112,12 +112,11 @@ def dodaj_wypozyczenie(request):
 @login_required
 def usun_wypozyczenie(request):
     id_wypozyczenia = request.GET.get('wypozyczenie_id','')
-    if id_wypozyczenia:
+    if id_wypozyczenia and id_wypozyczenia != 'brak wypozyczeń':
         query_delete = "DELETE from wypozyczenie WHERE wypozyczenie.id_wypozyczenia=%s"
         params = [id_wypozyczenia]
         wypozyczenia = write_operation(query_delete, params)
-
-   
+    
     query = "SELECT wypozyczenie.id_wypozyczenia, czytelnik.imie, czytelnik.nazwisko, tytul.tytul, autor.imie, autor.nazwisko, wypozyczenie.data_wypozyczenia, wypozyczenie.data_zwrotu from wypozyczenie JOIN czytelnik on czytelnik.id_czytelnika=wypozyczenie.id_czytelnika JOIN ksiazka on ksiazka.id_ksiazki=wypozyczenie.id_ksiazki JOIN tytul on tytul.id_tytulu=ksiazka.id_tytulu JOIN autor on autor.id_autora=tytul.id_autora"
     
     wypozyczenia = read_operation(query)
